expose tide testing trait publicly
